✨(font): glyph double-click → sidebar OpenedGlyph
font_panel.rs:
- glyph grid cell double_clicked() → build_opened_glyph() constructs
  OpenedGlyph from FreeTypeGlyph (true outline) or Mirx FontMetric (empty
  outline, approximate=true)
- push to state.opened_glyphs + state.items as SidebarItem::Glyph
- auto-select new item + switch to FontMode::Vector

model.rs:
- OpenedGlyph.bearing type (i8, i8) → (i16, i16) to unify FreeType (i16)
  and Mirx (i8 cast) bearing types

+++ src/image_viewer/model.rs
@@ -51,7 +51,7 @@ pub struct OpenedGlyph {
     pub codepoint: u32,
     pub char_repr: String,
     pub advance: u16,
-    pub bearing: (i8, i8),
+    pub bearing: (i16, i16),
     pub bbox: (i16, i16, i16, i16),
     pub outline: Vec<icu_lib::mirx::PathCmd>,
     pub outline_approximate: bool,
